+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/Art4/WP-Requests-PSR18-Adapter/compare/1.0.0-beta.1...HEAD)
 
+### Added
+
+- Add tests to improve test coverage
+
 ## [1.0.0-beta.1](https://github.com/Art4/WP-Requests-PSR18-Adapter/compare/1.0.0-beta...1.0.0-beta.1)
 
 ### Added